Move components around
This moves all `PublicServices::*` and `Shared::*` prefixed components to the top of the components folder.

Since only the LPDC module remains here that extra nesting is no longer needed.
