+++ app/components/bread-crumb.js
@@ -1,7 +1,7 @@
 import Component from '@glimmer/component';
 import { inject as service } from '@ember/service';
 
-export default class SharedBreadCrumbComponent extends Component {
+export default class BreadCrumbComponent extends Component {
   @service router;
 
   bread = [
