+++ app/components/concept-details-page.hbs
