+++ app/components/confirm-reopening-modal.js
@@ -2,7 +2,7 @@ import { action } from '@ember/object';
 import Component from '@glimmer/component';
 import { dropTask } from 'ember-concurrency';
 
-export default class PublicServicesConfirmReopeningModalComponent extends Component {
+export default class ConfirmReopeningModalComponent extends Component {
   @dropTask
   *reopen() {
     yield this.args.data.reopeningHandler();
