+++ app/components/details-page.hbs
