+++ app/components/footer-heading.js
@@ -1,6 +1,6 @@
 import Component from '@glimmer/component';
 
-export default class SharedFooterHeadingComponent extends Component {
+export default class FooterHeadingComponent extends Component {
   get hostname() {
     return document.location.hostname;
   }
