+++ app/components/unsaved-changes-modal.js
@@ -2,7 +2,7 @@ import { action } from '@ember/object';
 import Component from '@glimmer/component';
 import { dropTask } from 'ember-concurrency';
 
-export default class PublicServicesConfirmDiscardChangesModalComponent extends Component {
+export default class UnsavedChangesModalComponent extends Component {
   @dropTask
   *save() {
     yield this.args.data.saveHandler();
