+++ app/templates/application.hbs
@@ -60,7 +60,7 @@
                     Loket voor Lokale Besturen
                   </AuLinkExternal>
                 </li>
-                <Shared::BreadCrumb />
+                <BreadCrumb />
               </ul>
             </Group>
           </AuToolbar>
