+++ app/templates/legaal.hbs
@@ -7,7 +7,7 @@
     </div>
 
     <AuMainFooter>
-      <Shared::FooterHeading/>
+      <FooterHeading/>
       <AuContent @skin="small">
         <p>Uitgegeven door <AuLinkExternal href="https://www.vlaanderen.be/organisaties/administratieve-diensten-van-de-vlaamse-overheid/beleidsdomein-kanselarij-en-bestuur/agentschap-binnenlands-bestuur">Agentschap Binnenlands Bestuur</AuLinkExternal></p>
         <ul class="au-c-list-horizontal">
