+++ app/templates/login.hbs
@@ -36,7 +36,7 @@
     </main>
 
     <AuMainFooter>
-      <Shared::FooterHeading />
+      <FooterHeading />
       <AuContent @skin="small">
         <p>Uitgegeven door
           <a
