+++ app/templates/public-services/concept-details/content.hbs
@@ -1,6 +1,6 @@
 {{page-title "Inhoud"}}
 
-<PublicServices::ConceptDetailsPage
+<ConceptDetailsPage
   @concept={{@model.concept}}
   @formId="cd0b5eba-33c1-45d9-aed9-75194c3728d3"
   @tabName="au-c-rdf-form--content"
