+++ app/templates/public-services/concept-details/properties.hbs
@@ -1,6 +1,6 @@
 {{page-title "Eigenschappen"}}
 
-<PublicServices::ConceptDetailsPage
+<ConceptDetailsPage
   @concept={{@model.concept}}
   @formId="149a7247-0294-44a5-a281-0a4d3782b4fd"
   @tabName="au-c-rdf-form--properties"
