+++ app/templates/public-services/details/properties.hbs
@@ -1,6 +1,6 @@
 {{page-title "Eigenschappen"}}
 
-<PublicServices::DetailsPage
+<DetailsPage
   @publicService={{@model.publicService}}
   {{! We use `this.model` here since the value is used in the willDestroy method, and the `@model` version is undefined at that point.
       More information: https://github.com/emberjs/ember.js/issues/18987
